requirements.txt
requests==2.25.1
azure-cosmos==4.2.0